Updated Wave Spawner
New Wave Spawning system uses a seperate Wave class to allow for modified components such as rate, prefab, and count within each seperate wave. This is less ideal for the infinite game it currently is (so keep using old system if so) but otherwise its good for the future level based system we will develop.
Also added enums that allow devs to select between four different types of breaks between waves making it more modular and easier to control in the future.